// src/users/dto/bmi.dto.ts
export class BMIDto {
    name: string;
    age: number;
    gender: string;
    mobile: string;
    email: string;
    dob: string;
    height: number;
    weight: number;
    bmi: number;
    bmiCategory: string;
    minHealthyBmi: number;
    maxHealthyBmi: number;
    healthyWeightForHeight: number;
    gainToHealthyBmi: number;
    loseToHealthyBmi: number;
    ponderalIndex: number;
  }